requirements.txt
azure-iot-device==2.12.0
python-dotenv==1.0.0    